Updated index.html, level.js, and main.js
Updated on Jan. 23, 2021.

The index.html file is updated to include a checkbox element input for the bottom of the canvas. The checkbox is used to turn on and off the terrain grid map on the level image that is drawn on the canvas.

The level.js file now has 2 classes: a Level class and a LevelTerrainMap class. The Level class is used primarily for drawing a 2D map image graphic on canvas , and the LevelTerrainMap class is used for managing a 2D array that contains the terrain makeup of the level map image. The Level class constructs an object with a LevelTerrainMap field, and has a couple of methods to access and modify the LevelTerrainMap field. The LevelTerrainMap constructs an object with a 2D array field used to mark all terrain tiles on the map image, and in this version it can visually display the terrain tile grid map by having the user click on the "Grid Map" checkbox that's below the game window.

The main.js file is updated just to test creating a Level object and adding it as an entity to the gameengine class.